V3 adp decimal

+++ RAMS/Web/RAMMS.DTO/ResponseBO/FormV3DtlGridDTO.cs
@@ -24,7 +24,7 @@ namespace RAMMS.DTO.ResponseBO
         public string TransitTimeFrm { get; set; }
         public string TransitTimeTo { get; set; }
         public decimal? TransitTimeTotal { get; set; }
-        public int? Adp { get; set; }
+        public decimal? Adp { get; set; }
         public int? ModBy { get; set; }
         public DateTime? ModDt { get; set; }
         public int? CrBy { get; set; }
